Works better

+++ base/op.go
@@ -149,14 +149,15 @@ var LFOTypeNames = []string{
 	"SIN0", "SIN1", "RMP0", "RMP1", "COS0", "COS1",
 }
 
-var RampAmpValuesMap = map[int32]int32{
+var RampAmpValuesMap = map[int16]int32{
 	4096: 0,
 	2048: 1,
 	1024: 2,
 	512:  3,
 }
 
-var RampAmpValues = []int32{4096, 2048, 1024, 512}
+// FIXME: Correct order? Same as in disfv1.py, though. (20240713 handegar)
+var RampAmpValues = []int16{4096, 2048, 1024, 512}
 
 var SymbolEquivalents = map[int][]string{
 	0x00: {"SIN0_RATE",
